+++ tests/test_stiff.cc
@@ -15,10 +15,10 @@ int main () {
     printf("Solving system '%s' with method '%s'\n",
         sys.get_name(), sys.get_method());
     sys.solve_adaptive(tint, dt, "out", 1);
-    printf("%lu function evaluations\n", (unsigned long)sys.get_neval());
-    printf("%lu steps\n", (unsigned long)sys.get_nstep());
-    printf("%lu rejected steps\n", (unsigned long)sys.get_nrej());
-    printf("%lu ODE Jacobian evaluations\n", (unsigned long)sys.get_nJac());
+    printf("%lu function evaluations\n", sys.get_neval());
+    printf("%lu steps\n", sys.get_nstep());
+    printf("%lu rejected steps\n", sys.get_nrej());
+    printf("%lu ODE Jacobian evaluations\n", sys.get_nJac());
     printf("%lu Newton JLU updates (LU decompositions)\n", sys.get_newton()->get_nJLU());
     printf("%lu Newton LU solves (matrix equation solves)\n", sys.get_newton()->get_n_solve_LU());
